requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
langchain-openai
unstructured
beautifulsoup4
pydantic
requests
python-dotenv
beautifulsoup4
unstructured[pdf]
pdfminer.six
pypdf
cassio
astrapy
langchain-unstructured